Finally found a solution to the Choice Y / N Problem
GUIFD will now warn the user first time about piping echo letters through the entire alphabet to find the correct localized keys for Y / N. Due to the nature of Win-CMD, "Choice" command and other Microsoft stuff, this will create annyoing beep noises. But once the correct letter is found, the result will be saved.

Not optimal, but clearly usable. Could break, if the localized keys are not within the english alphabet.